requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
requests
strands-agents[ollama]>=1.0.0 
strands-agents-tools>=0.2.0
pytest